Added pbkdf2 (from https://defuse.ca), set up site skeleton

+++ onionimbus_node/INSTALL.sh
@@ -7,21 +7,27 @@ fi
 if [ -d /etc/nginx ] 
 then
   # nginx is installed -- hooray!
-  
-  # Create the directory for all of our configurations
-  mkdir /etc/onionimbus
-  
-  # Run this script every minute
-  crontab -l > mycron
-  echo "* * * * * /etc/onionimbus/everyminute.sh" >> mycron
-  crontab mycron
-  rm mycron
-  
-  # Deply configuration files
-  mv ./conf/nginx.conf /etc/onionimbus/nginx.conf
-  ln -s /etc/onionimbus/nginx.conf /etc/nginx/sites-enabled/onionimbus
-  # Move shell scripts
-  mv ./outside_webroot/everyminute.sh /etc/onionimbus/everyminute.sh
+  if [ -d /etc/tor ] 
+  then
+    # tor is installed -- hooray!
+    # Create the directory for all of our configurations
+    mkdir /etc/onionimbus
+    
+    # Run this script every minute
+    crontab -l > mycron
+    echo "* * * * * /etc/onionimbus/everyminute.sh" >> mycron
+    crontab mycron
+    rm mycron
+    
+    # Deply configuration files
+    mv ./conf/nginx.conf /etc/onionimbus/nginx.conf
+    ln -s /etc/onionimbus/nginx.conf /etc/nginx/sites-enabled/onionimbus
+    # Move shell scripts
+    mv ./outside_webroot/everyminute.sh /etc/onionimbus/everyminute.sh
+  else
+    echo "Please install Tor before you continue!" 1>&2
+    exit 1  
+  fi
 else 
   # nginx not installed
   echo "Please install nginx before you continue!" 1>&2
